release: SiriRemote 0.2.0 — secure audio and voice lifecycle (#5)
* release: prepare SiriRemote 0.2.0 with secure voice lifecycle

* test: synchronize resumed audio with published fixture frames

* test: derive release silence deadline from the actual queued tail

+++ mic/driver/srm_capture_test.c
@@ -167,9 +167,9 @@ static void inspect_shared_source(void)
         }
     }
     const double rms = sqrt(sumSquares / (sampleFrames * SRM_CHANNELS));
-    printf("srm_capture_test: source write=%llu active=%u rms=%.6f peak=%.6f\n",
+    printf("srm_capture_test: source write=%llu active=%llu rms=%.6f peak=%.6f\n",
            (unsigned long long)writeIndex,
-           atomic_load_explicit(&shared->producerActive, memory_order_acquire), rms, peak);
+           (unsigned long long)atomic_load_explicit(&shared->producerActive, memory_order_acquire), rms, peak);
     munmap(shared, sizeof(*shared));
 }
